Create new "Unit Test" BaseTest without properties
Also updated surefire and failsafe plugins to 2.19.1
