+++ src/test/java/com/esendex/java/sdk/integration/inboxservice/InboxServiceGetPagedMessagesIT.java
@@ -1,6 +1,6 @@
 package com.esendex.java.sdk.integration.inboxservice;
 
-import com.esendex.java.sdk.BaseTest;
+import com.esendex.java.sdk.integration.BaseTest;
 import esendex.sdk.java.EsendexException;
 import esendex.sdk.java.model.domain.response.InboxMessageCollectionResponse;
 import esendex.sdk.java.model.domain.response.InboxMessageResponse;
