+++ src/test/java/com/esendex/java/sdk/integration/inboxservice/InboxServiceMarkIT.java
@@ -1,6 +1,6 @@
 package com.esendex.java.sdk.integration.inboxservice;
 
-import com.esendex.java.sdk.BaseTest;
+import com.esendex.java.sdk.integration.BaseTest;
 import esendex.sdk.java.EsendexException;
 import esendex.sdk.java.model.domain.response.InboxMessageResponse;
 import esendex.sdk.java.service.InboxService;
